contrib/rexx: Compiler delint

git-svn-id: https://svn.aros.org/svn/aros/trunk@44353 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ contrib/rexx/lstring/time.c
@@ -167,7 +167,7 @@ Ltime( const PLstr timestr, char option )
 			gettimeofday(&tv,&tz);
 			sprintf(LSTR(*timestr), "%02d:%02d:%02d.%06ld",
 				tmdata->tm_hour, tmdata->tm_min,
-				tmdata->tm_sec, tv.tv_usec) ;
+				tmdata->tm_sec, (long)tv.tv_usec) ;
 #endif
 			break;
